sdk/requirements_dev.txt

traceloop-sdk>=0.47.5,<0.49.0
opentelemetry-instrumentation-fastapi>=0.50b0
pydantic>=2.8.2
setuptools
python-dotenv
langchain>=0.2.16,<0.4.0
langchain-openai
langgraph>=0.2.23,<0.5.0
langchain-ibm==0.3.5
langchain-community
ibm_watsonx_ai==1.1.26
crewai>=0.95.0,<0.131.0
crewai_tools
openai>=1.47.0
botocore
packaging
numpy==1.26.4
pymilvus
pymilvus[model]
flake8==7.1.1
litellm
git+ssh://git@github.ibm.com/agent-analytics/common.git@main
